TickRate was changed to a byte, no way it is going over 255. Fixed bug with changing TickRate after the loop has started. Client now properly sets the TickRate when connecting to a server.

+++ SS14.Shared/Interfaces/Timing/IGameTiming.cs
@@ -70,7 +70,7 @@ namespace SS14.Shared.Interfaces.Timing
         /// <summary>
         ///     The target ticks/second of the simulation.
         /// </summary>
-        int TickRate { get; set; }
+        byte TickRate { get; set; }
 
         /// <summary>
         ///     The length of a tick at the current TickRate. 1/TickRate.
